fix: move classes of data structures to separate folder
