+++ src/components/queue-page/queue-page.tsx
@@ -5,6 +5,7 @@ import { Input } from "../ui/input/input";
 import { Button } from "../ui/button/button";
 import { Circle } from "../ui/circle/circle";
 import { ElementStates } from "../../types/element-states";
+import { Queue } from "../../classes/Queue";
 
 export const QueuePage: React.FC = () => {
   
@@ -12,61 +13,6 @@ export const QueuePage: React.FC = () => {
     value: string;
     state: ElementStates;
   };
-
-  interface IQueue<T> {
-    enqueue: (item: T) => void;
-    dequeue: () => void;
-    peak: () => T | null;
-    getValues: () => T[];
-    getHead: () => number;
-    getTail: () => number;
-  }  
-  
-  class Queue<T> implements IQueue<T> {
-    private container: T[] = [];
-    head = 0;
-    tail = 0;
-    private readonly size: number = 0;
-    private length: number = 0;
-  
-    constructor(size: number) {
-      this.size = size;
-      this.container = Array(size);
-    }
-  
-    enqueue = (item: T) => {
-      if (this.length >= this.size) {
-        throw new Error("Maximum length exceeded");
-      }
-  
-      this.container[this.tail % this.size] = item;
-      this.length++;
-      this.tail++;
-    };
-  
-    dequeue = () => {
-      if (this.isEmpty()) {
-        throw new Error("No elements in the queue");
-      }
-  
-      delete this.container[this.head % this.size];
-      this.length--;
-      this.head++;
-    };
-  
-    peak = (): T | null => {
-      if (this.isEmpty()) {
-        throw new Error("No elements in the queue");
-      }
-  
-      return this.container[this.head];
-    };
-  
-    isEmpty = () => this.length === 0;
-    getValues = () => this.container;
-    getHead = () => this.head;
-    getTail = () => this.tail;
-  }
 
   const queueLength = 7;
   const maxLength = 4;
